test: split plot tests into finer-scoped files
